+++ bsp/lm3s8962/applications/startup.c
@@ -119,7 +119,7 @@ void rtthread_startup(void)
 #ifdef RT_USING_FINSH
     /* init finsh */
     finsh_system_init();
-#ifdef RT_USING_DEVICE
+#if !defined(RT_USING_LIBC) && defined(RT_USING_DEVICE)
     finsh_set_device("uart1");
 #endif
 #endif
