+++ components/dfs/SConscript
@@ -5,13 +5,9 @@ src = Split('''
 src/dfs.c
 src/dfs_file.c
 src/dfs_fs.c
-src/dfs_posix.c
 ''')
 cwd = GetCurrentDir()
 CPPPATH = [cwd + "/include"]
-
-if GetDepend('RT_USING_POSIX'):
-    src += ['src/dfs_poll.c', 'src/dfs_select.c']
 
 group = DefineGroup('Filesystem', src, depend = ['RT_USING_DFS'], CPPPATH = CPPPATH)
