+++ components/libc/compilers/common/SConscript
@@ -6,18 +6,15 @@ src   = []
 cwd   = GetCurrentDir()
 group = []
 CPPPATH = [cwd]
+CPPDEFINES = []
 
 if GetDepend('RT_USING_LIBC'):
     src += Glob('*.c')
-    if GetDepend('RT_USING_POSIX') == False:
-        SrcRemove(src, ['unistd.c', 'delay.c'])
 elif GetDepend('RT_LIBC_USING_TIME'):
     src += ['time.c']
 
 if rtconfig.CROSS_TOOL == 'keil':
-    CPPDEFINES = ['__CLK_TCK=RT_TICK_PER_SECOND']
-else:
-    CPPDEFINES = []
+    CPPDEFINES += ['__CLK_TCK=RT_TICK_PER_SECOND']
 
 group = DefineGroup('libc', src, depend = [], CPPPATH = CPPPATH, CPPDEFINES = CPPDEFINES)
