+++ components/libc/posix/getline/SConscript
@@ -6,7 +6,7 @@ cwd = GetCurrentDir()
 src = Glob('*.c') + Glob('*.cpp')
 CPPPATH = [cwd]
 
-group = DefineGroup('libc', src, 
+group = DefineGroup('POSIX', src,
     depend = ['RT_USING_LIBC', 'RT_USING_POSIX','RT_USING_POSIX_GETLINE'], 
     CPPPATH = CPPPATH)
