+++ components/libc/posix/mmap/SConscript
@@ -6,7 +6,7 @@ cwd = GetCurrentDir()
 src = Glob('*.c') + Glob('*.cpp')
 CPPPATH = [cwd]
 
-group = DefineGroup('libc', src, 
+group = DefineGroup('POSIX', src,
     depend = ['RT_USING_DFS', 'RT_USING_POSIX_MMAP'],
     CPPPATH = CPPPATH)
