+++ components/libc/posix/pthreads/SConscript
@@ -16,7 +16,7 @@ try:
 except :
     pass
 
-group = DefineGroup('libc', src, 
+group = DefineGroup('POSIX', src, 
     depend = ['RT_USING_PTHREADS', 'RT_USING_LIBC'], CPPPATH = CPPPATH, CPPDEFINES = CPPDEFINES)
 
 Return('group')
