+++ components/libc/posix/src/delay.c
@@ -9,6 +9,15 @@
  */
 #include <rtthread.h>
 #include <rthw.h>
+#define DBG_TAG    "POSIX.delay"
+#define DBG_LVL    DBG_INFO
+#include <rtdbg.h>
+
+RT_WEAK void rt_hw_us_delay(rt_uint32_t us)
+{
+    (void) us;
+    LOG_W("Please consider implementing rt_hw_us_delay() in another file.");
+}
 
 void msleep(unsigned int msecs)
 {
