+++ include/rtdebug.h
@@ -17,18 +17,7 @@
 #if !defined(RT_USING_DFS) || !defined(RT_USING_DFS_DEVFS)
 #error "POSIX poll/select, stdin need file system(RT_USING_DFS) and device file system(RT_USING_DFS_DEVFS)"
 #endif
-
-#if !defined(RT_USING_LIBC)
-#error "POSIX layer need standard C library(RT_USING_LIBC)"
-#endif
-
-#endif
-
-#ifdef RT_USING_POSIX_TERMIOS
-#if !defined(RT_USING_POSIX)
-#error "termios need POSIX layer(RT_USING_POSIX)"
-#endif
-#endif
+#endif /* RT_USING_POSIX */
 
 /* Using this macro to control all kernel debug features. */
 #ifdef RT_DEBUG
